+++ tooling/migration/README.md
@@ -1,0 +1,5 @@
+# Pre-requisites
+
+1. in order to use this code, you need to have a copy of `.env`
+2. inside your `.env`, put the `DATABASE_URL`
+3. go into the `index.ts` and update the site id to the site you want to migrate
